+++ lib/src/extensions/generic_type_x.dart
@@ -1,0 +1,24 @@
+//.title
+// ▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓
+//
+// Dart/Flutter (DF) Packages by DevCetra.com & contributors. The use of this
+// source code is governed by an MIT-style license described in the LICENSE
+// file located in this project's root directory.
+//
+// See: https://opensource.org/license/mit
+//
+// ▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓
+//.title~
+
+extension GenericTypeOnIterableX<T> on Iterable<T> {
+  /// Returns the generic type of the iterable.
+  Type get genericType => T;
+}
+
+extension GenericTypeOnMapX<K, V> on Map<K, V> {
+  /// Returns the generic type of the map's keys.
+  Type get genericTypeKey => K;
+
+  /// Returns the generic type of the map's values.
+  Type get genericTypeValue => V;
+}