+++ lib/src/extensions/map_filter_x.dart
@@ -10,14 +10,12 @@
 // ▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓
 //.title~
 
-extension MapWithDefaultOnMapExtension<K, V> on Map<K, V> {
+extension MapWithDefaultOnMapX<K, V> on Map<K, V> {
   /// Returns a new map with the same keys as this map but with the specified
   /// [defaultValue] for all values that are null. If [defaultValue] is null,
   /// it simply returns a copy of the original map.
   Map<K, dynamic> mapWithDefault(dynamic defaultValue) {
-    return defaultValue != null
-        ? map((k, v) => MapEntry(k, v ?? defaultValue))
-        : Map.of(this);
+    return defaultValue != null ? map((k, v) => MapEntry(k, v ?? defaultValue)) : Map.of(this);
   }
 
   /// Filters the map's entries based on a list of included values.