+++ lib/src/extensions/non_nulls_on_map_x.dart
@@ -10,7 +10,7 @@
 // ▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓
 //.title~
 
-extension NonNullsOnMapExtension<K, V> on Map<K?, V?> {
+extension NonNullsOnMapX<K, V> on Map<K?, V?> {
   /// Returns a new map with all non-null keys and values.
   Map<K, V> get nonNulls {
     return Map<K, V>.fromEntries(
@@ -21,24 +21,20 @@ extension NonNullsOnMapExtension<K, V> on Map<K?, V?> {
   }
 }
 
-extension NonNullKeysOnMapExtension<K, V> on Map<K?, V> {
+extension NonNullKeysOnMapX<K, V> on Map<K?, V> {
   /// Returns a new map with all non-null keys and values.
   Map<K, V> get nonNullKeys {
     return Map<K, V>.fromEntries(
-      entries
-          .where((e) => e.key != null)
-          .map((e) => MapEntry(e.key as K, e.value)),
+      entries.where((e) => e.key != null).map((e) => MapEntry(e.key as K, e.value)),
     );
   }
 }
 
-extension NonNullValuesOnMapExtension<K, V> on Map<K, V?> {
+extension NonNullValuesOnMapX<K, V> on Map<K, V?> {
   /// Returns a new map with all non-null keys and values.
   Map<K, V> get nonNullValues {
     return Map<K, V>.fromEntries(
-      entries
-          .where((e) => e.value != null)
-          .map((e) => MapEntry(e.key, e.value as V)),
+      entries.where((e) => e.value != null).map((e) => MapEntry(e.key, e.value as V)),
     );
   }
 }