+++ lib/src/extensions/unique_entries_x.dart
@@ -10,11 +10,7 @@
 // ▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓
 //.title~
 
-final _uniqueKeys = uniqueKeys;
-
-final _uniqueValues = uniqueValues;
-
-extension UniqueOnMapEntryIterableExtension<K, V> on Iterable<MapEntry<K, V>> {
+extension UniqueOnMapEntryIterableX<K, V> on Iterable<MapEntry<K, V>> {
   /// Returns only the unique entries in the iterable.
   List<MapEntry<K, V>> unique() => uniqueEntries(this);
 
@@ -25,6 +21,9 @@ extension UniqueOnMapEntryIterableExtension<K, V> on Iterable<MapEntry<K, V>> {
   List<MapEntry<K, V>> uniqueValues() => _uniqueValues(this);
 }
 
+final _uniqueKeys = uniqueKeys;
+final _uniqueValues = uniqueValues;
+
 // ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░
 
 List<MapEntry<K, V>> uniqueEntries<K, V>(Iterable<MapEntry<K, V>> entries) {
@@ -33,8 +32,7 @@ List<MapEntry<K, V>> uniqueEntries<K, V>(Iterable<MapEntry<K, V>> entries) {
   final unique = <MapEntry<K, V>>[];
 
   for (var entry in entries) {
-    if (!uniqueKeys.contains(entry.key) &&
-        !uniqueValues.contains(entry.value)) {
+    if (!uniqueKeys.contains(entry.key) && !uniqueValues.contains(entry.value)) {
       uniqueKeys.add(entry.key);
       uniqueValues.add(entry.value);
       unique.add(MapEntry(entry.key, entry.value));