+++ lib/src/extensions/unmodifiable_x.dart
@@ -1,0 +1,23 @@
+//.title
+// ▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓
+//
+// Dart/Flutter (DF) Packages by DevCetra.com & contributors. The use of this
+// source code is governed by an MIT-style license described in the LICENSE
+// file located in this project's root directory.
+//
+// See: https://opensource.org/license/mit
+//
+// ▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓
+//.title~
+
+extension UnmodifiableOnMapX<K, V> on Map<K, V> {
+  Map<K, V> get unmodifiable => Map.unmodifiable(this);
+}
+
+extension UnmodifiableOnListX<T> on Set<T> {
+  List<T> get unmodifiable => List.unmodifiable(this);
+}
+
+extension UnmodifiableOnSetX<T> on Set<T> {
+  Set<T> get unmodifiable => Set.unmodifiable(this);
+}